fix more windows warnings
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@38226 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 8ebb8f4779f7cda060e12ef4b8a8e36448164196

+++ Rec/Tr/TrackCheckers/src/TrackCheckerNT.cpp
@@ -1242,9 +1242,9 @@ StatusCode TrackCheckerNT::fillMCGeneralInfo(
     const MCVertex *mcvtx = mcPart->originVertex();
     if (mcvtx) {
       const Gaudi::XYZPoint &point = mcvtx->position();
-      _origVtxPos[nPart][0] = point.X() / mm;
-      _origVtxPos[nPart][1] = point.Y() / mm;
-      _origVtxPos[nPart][2] = point.Z() / mm;
+      _origVtxPos[nPart][0] = (float) (point.X() / mm);
+      _origVtxPos[nPart][1] = (float) (point.Y() / mm);
+      _origVtxPos[nPart][2] = (float) (point.Z() / mm);
     }
     // ok, write up to maxAssocTracks associated Tracks
     Track* track = revLink.first(mcPart);
